hugo: Update contribution guidelines
Refactor the contribution guidelines in the README and CONTRIBUTING
files.

Simplify the contribution guide in the README and move most of the
complex stuff into CONTRIBUTING.

Add an explicit commit message guidelines section to CONTRIBUTING.  Keep
all of the guidelines from Chris Beams except for the 72 character line
limit (we don't follow that, nor does the Go team).  Add three new
guidelines: package prefix in subject, references in body, and
encouragement of message body in general.

Add a new section to CONTRIBUTING on using Git Remotes.